Fix PillSheetModifiedHistoryChangedPillNumberAction
